+++ test/Proxy.test.js
@@ -3,8 +3,9 @@
 const path = require('path');
 const request = require('supertest');
 const express = require('express');
+const bodyParser = require('body-parser');
 const WebSocket = require('ws');
-const helper = require('./helper');
+const testServer = require('./helpers/test-server');
 const config = require('./fixtures/proxy-config/webpack.config');
 
 const WebSocketServer = WebSocket.Server;
@@ -83,7 +84,7 @@ describe('Proxy', () => {
 
     beforeAll((done) => {
       closeProxyServers = startProxyServers();
-      server = helper.start(
+      server = testServer.start(
         config,
         {
           contentBase,
@@ -95,7 +96,7 @@ describe('Proxy', () => {
     });
 
     afterAll((done) => {
-      helper.close(() => {
+      testServer.close(() => {
         closeProxyServers();
         done();
       });
@@ -139,7 +140,7 @@ describe('Proxy', () => {
 
     beforeAll((done) => {
       closeProxyServers = startProxyServers();
-      server = helper.start(
+      server = testServer.start(
         config,
         {
           contentBase,
@@ -151,7 +152,7 @@ describe('Proxy', () => {
     });
 
     afterAll((done) => {
-      helper.close(() => {
+      testServer.close(() => {
         closeProxyServers();
         done();
       });
@@ -169,7 +170,7 @@ describe('Proxy', () => {
 
     beforeAll((done) => {
       closeProxyServers = startProxyServers();
-      server = helper.start(
+      server = testServer.start(
         config,
         {
           contentBase,
@@ -181,7 +182,7 @@ describe('Proxy', () => {
     });
 
     afterAll((done) => {
-      helper.close(() => {
+      testServer.close(() => {
         closeProxyServers();
         done();
       });
@@ -210,7 +211,7 @@ describe('Proxy', () => {
         res.send('from proxy');
       });
       listener = proxy.listen(9000);
-      server = helper.start(
+      server = testServer.start(
         config,
         {
           contentBase,
@@ -225,7 +226,7 @@ describe('Proxy', () => {
     });
 
     afterAll((done) => {
-      helper.close(() => {
+      testServer.close(() => {
         listener.close();
         done();
       });
@@ -246,7 +247,7 @@ describe('Proxy', () => {
     let responseMessage;
 
     beforeAll((done) => {
-      helper.start(
+      testServer.start(
         config,
         {
           contentBase,
@@ -286,7 +287,103 @@ describe('Proxy', () => {
 
     afterAll((done) => {
       wsServer.close();
-      helper.close(done);
+      testServer.close(done);
+    });
+  });
+
+  describe('should support http methods', () => {
+    let server;
+    let req;
+    let listener;
+    const proxyTarget = {
+      target: 'http://localhost:9000',
+    };
+
+    beforeAll((done) => {
+      const proxy = express();
+
+      // Parse application/x-www-form-urlencoded
+      proxy.use(bodyParser.urlencoded({ extended: false }));
+
+      // Parse application/json
+      proxy.use(bodyParser.json());
+
+      proxy.get('/get', (proxyReq, res) => {
+        res.send('GET method from proxy');
+      });
+
+      proxy.head('/head', (proxyReq, res) => {
+        res.send('HEAD method from proxy');
+      });
+
+      proxy.post('/post-x-www-form-urlencoded', (proxyReq, res) => {
+        const id = proxyReq.body.id;
+
+        res.status(200).send(`POST method from proxy (id: ${id})`);
+      });
+
+      proxy.post('/post-application-json', (proxyReq, res) => {
+        const id = proxyReq.body.id;
+
+        res.status(200).send({ answer: `POST method from proxy (id: ${id})` });
+      });
+
+      proxy.delete('/delete', (proxyReq, res) => {
+        res.send('DELETE method from proxy');
+      });
+
+      listener = proxy.listen(9000);
+
+      server = testServer.start(
+        config,
+        {
+          contentBase,
+          proxy: {
+            '**': proxyTarget,
+          },
+        },
+        done
+      );
+      req = request(server.app);
+    });
+
+    afterAll((done) => {
+      testServer.close(() => {
+        listener.close();
+        done();
+      });
+    });
+
+    it('GET method', (done) => {
+      req.get('/get').expect(200, 'GET method from proxy', done);
+    });
+
+    it('HEAD method', (done) => {
+      req.head('/head').expect(200, done);
+    });
+
+    it('POST method (application/x-www-form-urlencoded)', (done) => {
+      req
+        .post('/post-x-www-form-urlencoded')
+        .send('id=1')
+        .expect(200, 'POST method from proxy (id: 1)', done);
+    });
+
+    it('POST method (application/json)', (done) => {
+      req
+        .post('/post-application-json')
+        .send({ id: '1' })
+        .set('Accept', 'application/json')
+        .expect('Content-Type', /json/)
+        .expect(
+          200,
+          JSON.stringify({ answer: 'POST method from proxy (id: 1)' }),
+          done
+        );
+    });
+
+    it('DELETE method', (done) => {
+      req.delete('/delete').expect(200, 'DELETE method from proxy', done);
     });
   });
 });