+++ test/options.test.js
@@ -120,8 +120,19 @@ describe('options', () => {
         failure: [false],
       },
       clientLogLevel: {
-        success: ['silent', 'info', 'error', 'warn', 'trace', 'debug'],
-        failure: ['whoops!', 'none', 'warning'],
+        success: [
+          'silent',
+          'info',
+          'error',
+          'warn',
+          'trace',
+          'debug',
+          // deprecated
+          'none',
+          // deprecated
+          'warning',
+        ],
+        failure: ['whoops!'],
       },
       compress: {
         success: [true],
@@ -359,6 +370,7 @@ describe('options', () => {
           {},
           'none',
           'errors-only',
+          'errors-warnings',
           'minimal',
           'normal',
           'verbose',